Replace usage of long long with u64/s64

+++ src/database-dummy.cpp
@@ -151,7 +151,7 @@ MapBlock* Database_Dummy::loadBlock(v3s16 blockpos)
 
 void Database_Dummy::listAllLoadableBlocks(std::list<v3s16> &dst)
 {
-	for(std::map<unsigned long long, std::string>::iterator x = m_database.begin(); x != m_database.end(); ++x)
+	for(std::map<u64, std::string>::iterator x = m_database.begin(); x != m_database.end(); ++x)
 	{
 		v3s16 p = getIntegerAsBlock(x->first);
 		//dstream<<"block_i="<<block_i<<" p="<<PP(p)<<std::endl;
