+++ src/util/serialize.cpp
@@ -22,6 +22,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "porting.h"
 #include "util/string.h"
 #include "../exceptions.h"
+#include "../irrlichttypes.h"
 
 #include <sstream>
 #include <iomanip>
@@ -428,7 +429,7 @@ bool serializeStructToString(std::string *outstr,
 					bufpos += PADDING(bufpos, u64);
 					nprinted = snprintf(sbuf + pos, sbuflen,
 								is_unsigned ? "%llu, " : "%lli, ",
-								(unsigned long long)*((u64 *)bufpos));
+								*((u64 *)bufpos));
 					bufpos += sizeof(u64);
 				}
 				break;
